requirements-dev.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.2
django-admin-rangefilter==0.8.1
django-bootstrap-form==3.4
django-recaptcha==2.0.6
django-semantic-ui==1.2.3
django-summernote==0.8.11.6
idna==2.10
importlib-metadata==4.2.0
jsonconfig-tool==1.5
keyring==23.0.1
messages==0.5.0
Pillow==8.2.0
python-box==5.3.0
python-decouple==3.4
pytz==2021.1
pywin32-ctypes==0.2.0
requests==2.25.1
sqlparse==0.4.1
static3==0.7.0
urllib3==1.26.5
validus==0.3.0
zipp==3.4.1
